Renamed class package and added run description
